Moved the UISearchBar extension into its own file in the Helpers directory
