+++ Resfeber/Helpers/UISearchBar+Extension.swift
@@ -1,0 +1,21 @@
+//
+//  UISearchBar+Extension.swift
+//  Resfeber
+//
+//  Created by David Wright on 11/9/20.
+//  Copyright © 2020 Spencer Curtis. All rights reserved.
+//
+
+import UIKit
+
+extension UISearchBar {
+    func clearBackgroundColor() {
+        guard let UISearchBarBackground: AnyClass = NSClassFromString("UISearchBarBackground") else { return }
+
+        for view in subviews {
+            for subview in view.subviews where subview.isKind(of: UISearchBarBackground) {
+                subview.alpha = 0
+            }
+        }
+    }
+}
