+++ Resfeber/View Controllers/ExploreViewController.swift
@@ -101,15 +101,3 @@ extension ExploreViewController: UISearchBarDelegate {
         performQuery(with: searchBar.text)
     }
 }
-
-extension UISearchBar {
-    func clearBackgroundColor() {
-        guard let UISearchBarBackground: AnyClass = NSClassFromString("UISearchBarBackground") else { return }
-
-        for view in subviews {
-            for subview in view.subviews where subview.isKind(of: UISearchBarBackground) {
-                subview.alpha = 0
-            }
-        }
-    }
-}
